requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.2.1
blinker==1.9.0
browser-cookie3==0.19.1
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.110.2
Flask==3.1.0
gmssl==3.2.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.0
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.5
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.2
orjson==3.10.15
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.0
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pydantic_core==2.18.1
pyfiglet==1.0.0
Pygments==2.17.2
PyMySQL==1.1.1
pypng==0.20220715.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pywebio==1.8.3
pywebio_battery==0.6.0
PyYAML==6.0.1
qrcode==7.4.2
requests==2.32.3
rich==13.7.1
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.37.2
tenacity==9.0.0
tornado==6.4
typing_extensions==4.11.0
ua-parser==0.18.0
ujson==5.10.0
urllib3==2.3.0
user-agents==2.2.0
uvicorn==0.29.0
watchfiles==1.0.4
websockets==12.0
Werkzeug==3.1.3
